# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notification Service API",
        "description": "Микросервис массовых уведомлений SMS/Email с приоритизацией и дедупликацией",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/notifications/batch": {
            "post": {
                "tags": [
                    "Batches"
                ],
                "summary": "Запустить массовую рассылку",
                "description": "Создаёт батч и ставит уведомления в очередь. Поддерживает заголовок Idempotency-Key.",
                "operationId": "d348cf283d0bdab92aa506120e1c03c1",
                "parameters": [
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "description": "UUID для идемпотентности. Повторный запрос с тем же ключом вернёт 200 с кэшированным ответом.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "channel",
                                    "priority",
                                    "message_text",
                                    "recipient_ids"
                                ],
                                "properties": {
                                    "channel": {
                                        "type": "string",
                                        "example": "sms",
                                        "enum": [
                                            "sms",
                                            "email"
                                        ]
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "transactional",
                                        "enum": [
                                            "transactional",
                                            "marketing"
                                        ]
                                    },
                                    "message_text": {
                                        "type": "string",
                                        "example": "Ваш код доступа: 1234",
                                        "maxLength": 1000
                                    },
                                    "recipient_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "maxItems": 1000,
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Рассылка поставлена в очередь",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/NotificationBatch"
                                        },
                                        "meta": {
                                            "properties": {
                                                "idempotent": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Повторный запрос — идемпотентный ответ"
                    },
                    "409": {
                        "description": "Запрос с этим Idempotency-Key уже обрабатывается"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/v1/notifications/batches/{batchId}": {
            "get": {
                "tags": [
                    "Batches"
                ],
                "summary": "Получить статус батча",
                "operationId": "affb0dc54ce95a3f38d725c73cee5143",
                "parameters": [
                    {
                        "name": "batchId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус батча",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/NotificationBatch"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Батч не найден"
                    }
                }
            }
        },
        "/api/v1/recipients/{recipientId}/notifications": {
            "get": {
                "tags": [
                    "Recipients"
                ],
                "summary": "История уведомлений получателя",
                "operationId": "aab6e6c1ff537bdf406f62887b120608",
                "parameters": [
                    {
                        "name": "recipientId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "queued",
                                "sent",
                                "delivered",
                                "discarded"
                            ]
                        }
                    },
                    {
                        "name": "channel",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "sms",
                                "email"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "История уведомлений",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Notification"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Получатель не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации параметров"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "NotificationBatch": {
                "properties": {
                    "batch_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "channel": {
                        "type": "string",
                        "enum": [
                            "sms",
                            "email"
                        ]
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "transactional",
                            "marketing"
                        ]
                    },
                    "message_text": {
                        "type": "string"
                    },
                    "total_count": {
                        "type": "integer"
                    },
                    "queued_count": {
                        "type": "integer"
                    },
                    "sent_count": {
                        "type": "integer"
                    },
                    "delivered_count": {
                        "type": "integer"
                    },
                    "discarded_count": {
                        "type": "integer"
                    },
                    "progress_percent": {
                        "type": "number"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Notification": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "batch_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "channel": {
                        "type": "string",
                        "enum": [
                            "sms",
                            "email"
                        ]
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "transactional",
                            "marketing"
                        ]
                    },
                    "message_text": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "queued",
                            "sent",
                            "delivered",
                            "discarded"
                        ]
                    },
                    "provider_message_id": {
                        "type": "string",
                        "nullable": true
                    },
                    "error_message": {
                        "type": "string",
                        "nullable": true
                    },
                    "attempts": {
                        "type": "integer"
                    },
                    "sent_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "delivered_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "discarded_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Batches",
            "description": "Управление массовыми рассылками"
        },
        {
            "name": "Recipients",
            "description": "История уведомлений получателей"
        }
    ]
}